Added resource based authorization

+++ src/Restaurants.Application/Dishes/Commands/DeleteAllDishesForRestaurant/DeleteAllDishesForRestaurantCommandHandler.cs
@@ -1,14 +1,17 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Authorization;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteAllDishesForRestaurant;
 
 public class DeleteAllDishesForRestaurantCommandHandler(ILogger<DeleteAllDishesForRestaurantCommandHandler> logger,
     IRestaurantsRepository restaurantsRepository,
-    IDishesRepository dishesRepository) : IRequestHandler<DeleteAllDishesForRestaurantCommand>
+    IDishesRepository dishesRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteAllDishesForRestaurantCommand>
 {
     public async Task Handle(DeleteAllDishesForRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -17,6 +20,11 @@ public class DeleteAllDishesForRestaurantCommandHandler(ILogger<DeleteAllDishesF
         if (restaurant is null)
         {
             throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+        }
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
+        {
+            throw new ForbidException();
         }
 
         await dishesRepository.DeleteAsync(restaurant.Dishes);
